+++ pkg/providers/property/testdata/TestDataIncludeActivation/no_include_id.tf
@@ -1,0 +1,10 @@
+provider "akamai" {
+  edgerc = "../../test/edgerc"
+}
+
+
+data "akamai_include_activation" "test" {
+  contract_id = "contract_123"
+  group_id    = "group_321"
+  network     = "STAGING"
+}